+++ client/src/Components/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Route, Switch } from "react-router-dom";
 
-import UploadImage from "./UploadImage";
+import UploadImage from "./Pages/UploadImage";
 import Thankyou from "./Pages/ThankYou";
 
 export default function index(props) {
